pyproject.toml
[project]
name = "featherharness"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipython>=9.12.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
]
